+++ fn.go
@@ -26,7 +26,7 @@ func (f *Function) RunFunction(_ context.Context, req *fnv1beta1.RunFunctionRequ
 		composed       *composite.Composition
 		input          inp.Input
 		vpcs           map[string]fnc.Vpc = make(map[string]fnc.Vpc)
-		search         []*inp.RemoteVpc   = make([]*inp.RemoteVpc, 0)
+		search         []inp.RemoteVpc    = make([]inp.RemoteVpc, 0)
 		region         string
 		providerConfig string
 	)
@@ -77,7 +77,7 @@ func (f *Function) RunFunction(_ context.Context, req *fnv1beta1.RunFunctionRequ
 	}
 	f.log.Info("ProviderConfig", "pc", providerConfig)
 
-	if err = f.getValueInto(oxr.Resource, input.Spec.VpcNameRef, region, providerConfig, groupTag, search); err != nil {
+	if err = f.getValueInto(oxr.Resource, input.Spec.VpcNameRef, region, providerConfig, groupTag, &search); err != nil {
 		f.log.Info("cannot get VPC name from input", "error", err)
 		response.Fatal(rsp, errors.Wrap(err, "cannot get VPC name from input"))
 		return rsp, nil
@@ -85,8 +85,8 @@ func (f *Function) RunFunction(_ context.Context, req *fnv1beta1.RunFunctionRequ
 
 	for _, n := range search {
 		var vpc fnc.Vpc
-		if vpc, err = f.ReadVpc(n); err != nil {
-			f.log.Info("cannot read VPC", "error", err, "name", n.Name)
+		if vpc, err = f.ReadVpc(&n); err != nil {
+			f.log.Info("cannot read VPC", "error", err, "name", n.Name, "region", n.Region, "providerConfig", n.ProviderConfig)
 			continue
 		}
 
@@ -114,7 +114,7 @@ func (f *Function) RunFunction(_ context.Context, req *fnv1beta1.RunFunctionRequ
 }
 
 // get array from paved
-func (f *Function) getValueInto(req runtime.Object, ref, region, providerConfig, groupBy string, value []*inp.RemoteVpc) (err error) {
+func (f *Function) getValueInto(req runtime.Object, ref, region, providerConfig, groupBy string, value *[]inp.RemoteVpc) (err error) {
 	var paved *fieldpath.Paved
 	if paved, err = fieldpath.PaveObject(req); err != nil {
 		return
@@ -122,18 +122,18 @@ func (f *Function) getValueInto(req runtime.Object, ref, region, providerConfig,
 
 	var s string
 	if s, err = paved.GetString(ref); err != nil {
-		err = paved.GetValueInto(ref, value)
-		for i := range value {
-			if value[i].Region == "" {
-				value[i].Region = region
+		err = paved.GetValueInto(ref, &value)
+		for i := range *value {
+			if (*value)[i].Region == "" {
+				(*value)[i].Region = region
 			}
 
-			if value[i].ProviderConfig == "" {
-				value[i].ProviderConfig = providerConfig
+			if (*value)[i].ProviderConfig == "" {
+				(*value)[i].ProviderConfig = providerConfig
 			}
 
-			if value[i].GroupBy == "" {
-				value[i].GroupBy = groupBy
+			if (*value)[i].GroupBy == "" {
+				(*value)[i].GroupBy = groupBy
 			}
 		}
 		return
@@ -144,7 +144,7 @@ func (f *Function) getValueInto(req runtime.Object, ref, region, providerConfig,
 		Region:         region,
 		ProviderConfig: providerConfig,
 	}
-	_ = append(value, &input)
+	*value = append(*value, input)
 	return
 }