+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.4] - 2024-06-20
+
+- handle service endpoints locally
+
 ## [0.2.3] - 2024-06-17
 
 - Fix missing region and provider config
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.2] - 2024-06-17
 
 - Remove primary cidr block from list of additional cidrs to prevent duplication
+- Fix missing region and provider config
 
 ## [0.2.1] - 2024-06-17
 
@@ -27,7 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of function-network-discovery
 
-[Unreleased]: https://github.com/giantswarm/crossplane-fn-network-discovery/compare/v0.2.3...HEAD
+[Unreleased]: https://github.com/giantswarm/crossplane-fn-network-discovery/compare/v0.2.4...HEAD
+[0.2.4]: https://github.com/giantswarm/crossplane-fn-network-discovery/compare/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/giantswarm/crossplane-fn-network-discovery/compare/v0.2.1...v0.2.3
 [0.2.2]: https://github.com/giantswarm/crossplane-fn-network-discovery/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/giantswarm/crossplane-fn-network-discovery/compare/v0.2.0...v0.2.1